Small refactoring
1) Add yocto_httpserver()
2) Take the HOSTNAME from the local.conf MACHINE name.

+++ debian.do-all.sh
@@ -4,6 +4,22 @@ PROGNAME=${BASH_SOURCE[0]}
 DIRNAME=$(dirname ${PROGNAME})
 [ $(basename -- $BASH_SOURCE) == $(basename -- $0) ] && EXIT="exit" || EXIT="return"
 
+function yocto_httpserver() {
+
+	pushd .
+
+	cd ${BUILDDIR}/../
+
+	source setup-environment ${BUILD_DIR}
+	bitbake ${PACKAGES} package-index
+	cd tmp/deploy/deb
+	python -m SimpleHTTPServer 5678 &
+	export PID=$!
+
+	popd
+
+}
+
 CONF=${DIRNAME}/compulab.inc
 [[ ! -e ${CONF} ]] && ${EXIT} 3
 . ${CONF}
@@ -12,21 +28,18 @@ CONF=${DIRNAME}/../local/local.conf
 [[ ! -e ${CONF} ]] && ${EXIT} 2
 . ${CONF}
 
+DEBIAN_CONF=debian.config.inc
+if [[ ! -e ${DIRNAME}/${DEBIAN_CONF} ]];then
+cat << eof | tee -a ${DIRNAME}/${DEBIAN_CONF}
+HOSTNAME=${MACHINE}
+eof
+fi
+
 BUILDDIR=$(dirname $(dirname ${DEPLOY_DIR})) 
 BUILD_DIR=$(basename ${BUILDDIR})
 
-pushd .
-
-cd ${BUILDDIR}/../
-
-source setup-environment ${BUILD_DIR}
-bitbake ${PACKAGES} package-index
-cd tmp/deploy/deb
-python -m SimpleHTTPServer 5678 &
-PID=$!
-
-popd
+yocto_httpserver
 
 ${DIRNAME}/debian.sh
 
-kill -9 ${PID}
+[[ -n ${PID} ]] && kill -9 ${PID}
